requirements.txt
playwright
moviepy
google-generativeai
firebase-admin
requests
python-dotenv
pytest
foundry-local-sdk
openai
edge-tts
google-auth-oauthlib
google-api-python-client
Pillow
